refactor: extract compose domain logic from CLI to orchestrator
Move compose_features.rs (feature resolution, combined Dockerfile
assembly) and compose_up.rs (compose override, lifecycle, container
management) into cella-orchestrator.

CLI retains thin wrappers: CliComposeUpHooks implements
ComposeUpHooks for daemon-specific operations (registration, agent
launch, post-create setup). Same proven pattern as PruneHooks.

Also removes dead lifecycle.rs wrapper and unused helper functions
from up/mod.rs.

Ref: #15
